+++ Model/Downloads/DownloadTasksPublisher.swift
@@ -31,7 +31,7 @@ final class DownloadTasksPublisher {
         }
     }
 
-    func updateFor(uuid: UUID, downloaded: Int64, total: Int64) {
+    func updateFor(uuid: UUID, downloaded: UInt, total: UInt) {
         if let state = states[uuid] {
             states[uuid] = state.updatedWith(downloaded: downloaded, total: total)
         } else {
@@ -48,7 +48,7 @@ final class DownloadTasksPublisher {
     }
     
     #if os(macOS)
-    func offsetFor(uuid: UUID) -> Int64? {
+    func offsetFor(uuid: UUID) -> UInt? {
         states[uuid]?.downloaded
     }
     #endif
